Add return value to silence a warning
pthreads can return a value, and the joiner can optionally capture this
value at an address they pass in. The code doesn't use the return value
so it doesn't matter, but not having a return in the code can lead to
junk getting sent back. New compilers warn for this, so fixing it to silence
the warning in noetic